+++ docs/changelog.rst
@@ -6,6 +6,14 @@ What's new
 .. contents:: Table of Contents
    :depth: 2
 
+v0.2.8 (July 2019)
+------------------
+
+**Enhancements**
+
+a. :py:func:`pingouin.anova` now three or more between factors (with statsmodels). One-way ANOVA and balanced two-way ANOVA are computed in pure Pingouin (Python + Pandas) style, while ANOVA with three or more factors, or unbalanced two-way ANOVA are computed using statsmodels.
+b. Added ss_type to :py:func:`pingouin.anova` for unbalanced N-way design.
+
 v0.2.7 (June 2019)
 ------------------